+++ jest/setup.js
@@ -1,9 +1,9 @@
-import fs from 'fs';
-import path from 'path';
 import 'react-native-gesture-handler/jestSetup';
-import _ from 'underscore';
+import * as reanimatedJestUtils from 'react-native-reanimated/lib/reanimated2/jestUtils';
+import setupMockImages from './setupMockImages';
 
-require('react-native-reanimated/lib/reanimated2/jestUtils').setUpTests();
+setupMockImages();
+reanimatedJestUtils.setUpTests();
 
 // This mock is required as per setup instructions for react-navigation testing
 // https://reactnavigation.org/docs/testing/#mocking-native-modules
@@ -19,26 +19,3 @@ jest.spyOn(console, 'debug').mockImplementation((...params) => {
     // eslint-disable-next-line no-console
     console.log('DEBUG', ...params);
 });
-
-/**
- * @param {String} imagePath
- */
-function mockImages(imagePath) {
-    const imageFilenames = fs.readdirSync(path.resolve(__dirname, `../assets/${imagePath}/`));
-    // eslint-disable-next-line rulesdir/prefer-early-return
-    _.each(imageFilenames, (fileName) => {
-        if (/\.svg/.test(fileName)) {
-            jest.mock(`../assets/${imagePath}/${fileName}`, () => () => '');
-        }
-    });
-}
-
-// We are mocking all images so that Icons and other assets cannot break tests. In the testing environment, importing things like .svg
-// directly will lead to undefined variables instead of a component or string (which is what React expects). Loading these assets is
-// not required as the test environment does not actually render any UI anywhere and just needs them to noop so the test renderer
-// (which is a virtual implemented DOM) can do it's thing.
-mockImages('images');
-mockImages('images/avatars');
-mockImages('images/bankicons');
-mockImages('images/product-illustrations');
-jest.mock('../src/components/Icon/Expensicons');
